[bug] workaround wrong qtplugins path rewritten in the libs
this is a bug in macholib, there's a missing letter in the plugins path, so
unless we fix this they cannot be loaded from the bundle.

See:
https://bitbucket.org/pydica/pyside-setup/commits/4b8be97e5a00b577fe30ce9aa7e5723ff2a66f94

Quoting from http://code.activestate.com/lists/pythonmac-sig/23278/:

"""
The problem might be this line:
@rpath/Contents/mageformats/libqtiff.dylib

The "i" from "imageformats" is missing!

This _might_ be related to the unusual case that "libqtiff" has no path
at all, or something else
is funny, and we end up with a name that will not be found at all.
Then the loader finds the plugin in the installed Qt, which causes it to
load everything
again from there.
"""